Export Mode from ModeSwitcher and make new folder
